remove unused endpoint components and move endpoint selector component
